[refactor] Move image.py from misc/ to tools/ and remove it from top level package
