+++ python/taichi/misc/__init__.py
@@ -1,5 +1,4 @@
 from .error import *
-from .image import *
 from .util import *
 
 __all__ = [s for s in dir() if not s.startswith('_')]
