+++ python/taichi/tools/__init__.py
@@ -1,5 +1,6 @@
+from .image import imdisplay, imread, imresize, imshow, imwrite
 from .np2ply import PLYWriter
 from .patterns import taichi_logo
 from .video import VideoManager
 
-__all__ = ['PLYWriter', 'taichi_logo', 'VideoManager']
+__all__ = ['PLYWriter', 'taichi_logo', 'VideoManager', 'imdisplay', 'imread', 'imresize', 'imshow', 'imwrite']
