+++ python/taichi/tools/image.py
@@ -133,12 +133,3 @@ def imshow(img, window_name='imshow'):
 
             gui.set_image(img)
             gui.show()
-
-
-__all__ = [
-    'imshow',
-    'imread',
-    'imwrite',
-    'imresize',
-    'imdisplay',
-]
