+++ python/taichi/tools/video.py
@@ -2,7 +2,7 @@ import os
 import shutil
 
 from taichi.core import get_os_name
-from taichi.misc.image import imwrite
+from taichi.tools.image import imwrite
 
 FRAME_FN_TEMPLATE = '%06d.png'
 FRAME_DIR = 'frames'
